【修改人】liushaofeng 【修改工程】tangchao 【修改记录】添加首页中当前会员的基本信息显示

+++ src/tk/shaofeng/tangchao/dao/RechargeDAO.java
@@ -64,7 +64,7 @@ public class RechargeDAO extends AbstractTCDAO
 	 * @return recharge records
 	 */
 	@SuppressWarnings("unchecked")
-	public List<RechargeModel> findByCardIdDesc(long cardId)
+	public List<RechargeModel> findByCardIdDesc(int cardId)
 	{
 		String hql = "from RechargeModel model where model.cardId = ? order by model.rechageDate desc";
 		Session session = getSession();
